pyproject.toml
[project]
name = "movie-viewer"
version = "0.3.0"
description = "Watch downloaded movies and shows easly"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.10.1",
    "nuitka>=4.0.2",
    "pillow>=11.2.1",
    "pyqt6>=6.9.0",
    "python-dotenv>=1.1.0",
    "qt-material>=2.17",
    "requests>=2.32.3",
]
